// src/app/shared/enum/header-grid-relatorio-generico.enum.ts
export enum HeaderGridRelatorioGenerico {

    tipo = "Tipo",
    fase = "Fase",
    complemento1 = "Complemento 1",
    complemento2 = "Complemento 2",
    dataInicial = "Data Inicial da Fase",
    dataFinal = "Data Final da Fase",
    naoExiste = "Não Existe",
    ultimaFase = "Última Fase",
    acoes = "Ações"

}